docker-lab/api/requirements.txt
flask
prometheus-flask-exporter
psycopg2-binary
